toggle isActive on click

+++ server/app/controllers/api/frontend/DeliveryTimesController.php
@@ -1,5 +1,6 @@
 <?php namespace App\Controllers\Api\Frontend;
 
+use Validator;
 use Input;
 use Request;
 
@@ -8,20 +9,28 @@ use App\Models\DeliveryTimeModel;
 class DeliveryTimesController extends ApiController
 {
 
-	public function __construct() {
-		$this->checkAuthentification();
-	}
-
 	public function create()
 	{
 		$this->loadStoreModel();
+		$this->checkAuthentification();
 
 		if ($this->hasErrorOccured()) {
 			return $this->respondWithError();
 		}
 
+		// check input
 		$input = Input::json();
+		$rules = array(
+			'dayOfWeek'	=> 'numeric|required|between:0,6'
+			);
 
+		$validator = Validator::make(get_object_vars($input), $rules);
+
+		if ($validator->fails()) {
+			return $this->respondWithStatus(400, $validator->messages());
+		}
+
+		// create new deliveryTimeModel
 		$deliveryTimeModel = new DeliveryTimeModel(array(
 			'store_model_id' => $this->storeModel->id,
 			'startMinutes' => 0,
@@ -29,24 +38,50 @@ class DeliveryTimesController extends ApiController
 			'dayOfWeek' => $input->dayOfWeek
 			));
 
+		// save
 		$deliveryTimeModel->save();
 
+		// return as json
 		return $deliveryTimeModel->toJson(JSON_NUMERIC_CHECK);
 	}
 
 	public function update()
 	{
-		$input = Input::json();
-		$id = Request::segment(6);
+		$this->checkAuthentification();
 
+		if ($this->hasErrorOccured()) {
+			return $this->respondWithError();
+		}
+
+		// check input
+		$input = Input::json();
+		$rules = array(
+			'startMinutes'	=> 'numeric|required|between:0,1438',
+			'endMinutes'	=> 'numeric|required|between:0,1439'
+			);
+
+		$validator = Validator::make(get_object_vars($input), $rules);
+
+		if ($validator->fails()) {
+			return $this->respondWithStatus(400, $validator->messages());
+		}
+
+		if ($input->startMinutes >= $input->endMinutes) {
+			return $this->respondWithStatus(400, 'endMinutes must be bigger then startMinutes');
+		}
+
+		// fetch deliveryTimeModel
+		$id = Request::segment(6);
 		$deliveryTimeModel = DeliveryTimeModel::find($id);
 
+		// verify owner
 		$this->checkBelongsToThisStore($deliveryTimeModel->storeModel->id);
 
 		if ($this->hasErrorOccured()) {
 			return $this->respondWithError();
 		}
 
+		// update item
 		$deliveryTimeModel->startMinutes = $input->startMinutes;
 		$deliveryTimeModel->endMinutes = $input->endMinutes;
 
@@ -55,9 +90,22 @@ class DeliveryTimesController extends ApiController
 
 	public function destroy()
 	{
-		$id = Request::segment(6);
+		$this->checkAuthentification();
+		
+		if ($this->hasErrorOccured()) {
+			return $this->respondWithError();
+		}
 
+		// fetch deliveryTimeModel
+		$id = Request::segment(6);
 		$deliveryTimeModel = DeliveryTimeModel::find($id);
+
+		// verify owner
+		$this->checkBelongsToThisStore($deliveryTimeModel->storeModel->id);
+
+		if ($this->hasErrorOccured()) {
+			return $this->respondWithError();
+		}
 
 		$deliveryTimeModel->delete();
 	}
